flutterapi/requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
cmd2==2.3.3
component==0.0.1
cycler==0.10.0
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.4.0
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.3
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
Mypackage==0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.0
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.1
pyperclip==1.8.2
pyreadline3==3.3
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
